tests: fix the directory fixture, add autologging for c-lightning

+++ tests/fixtures.py
@@ -23,8 +23,11 @@ def test_base_dir():
 
     yield directory
 
-    if os.listdir(directory) == []:
+    content = os.listdir(directory)
+    if content == []:
         shutil.rmtree(directory)
+    else:
+        print(f"Leaving base dir '{directory}' as it still contains {content}")
 
 
 @pytest.fixture
@@ -41,15 +44,51 @@ def directory(request, test_base_dir, test_name):
                              "{}_{}".format(test_name, __attempts[test_name]))
     request.node.has_errors = False
 
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+
     yield directory
 
-    # FIXME: use lightningd's teardown checks for errors
-    try:
-        shutil.rmtree(directory)
-    except Exception:
-        files = [os.path.join(dp, f) for dp, _, fn in os.walk(directory) for f in fn]
-        print("Directory still contains files:", files)
-        raise
+    # This uses the status set in conftest.pytest_runtest_makereport to
+    # determine whether we succeeded or failed. Outcome can be None if the
+    # failure occurs during the setup phase, hence the use to getattr instead
+    # of accessing it directly.
+    rep_call = getattr(request.node, 'rep_call', None)
+    outcome = 'passed' if rep_call is None else rep_call.outcome
+    failed = not outcome or request.node.has_errors or outcome != 'passed'
+
+    if not failed:
+        try:
+            shutil.rmtree(directory)
+        except Exception:
+            files = [os.path.join(dp, f) for dp, _, fn in os.walk(directory) for f in fn]
+            print("Directory still contains files:", files)
+            raise
+    else:
+        print(f"Test failed, leaving directory '{directory}' intact")
+
+
+@pytest.fixture(autouse=True)
+def setup_logging():
+    """Enable logging before a test, and remove all handlers afterwards.
+
+    This "fixes" the issue with pytest swapping out sys.stdout and sys.stderr
+    in order to capture the output, but then doesn't wait for the handlers to
+    terminate before closing the buffers. It just iterates through all
+    loggers, and removes any handlers that might be pointing at sys.stdout or
+    sys.stderr.
+
+    """
+    if TEST_DEBUG:
+        logging.basicConfig(level=logging.DEBUG, stream=sys.stdout)
+
+    yield
+
+    loggers = [logging.getLogger()] + list(logging.Logger.manager.loggerDict.values())
+    for logger in loggers:
+        handlers = getattr(logger, 'handlers', [])
+        for handler in handlers:
+            logger.removeHandler(handler)
 
 
 @pytest.fixture
@@ -94,10 +133,9 @@ def revaultd_stakeholder(bitcoind, directory):
     revaultd.cleanup()
 
 
-# FIXME: Fix the directory fixture to conserve dir on failure
 @pytest.fixture
-def revaultd_manager(bitcoind, test_base_dir):
-    datadir = os.path.join(test_base_dir, "revaultd")
+def revaultd_manager(bitcoind, directory):
+    datadir = os.path.join(directory, "revaultd")
     os.makedirs(datadir, exist_ok=True)
     stakeholders, managers = get_participants(2, 3)
 
@@ -115,8 +153,8 @@ def revaultd_manager(bitcoind, test_base_dir):
 
 
 @pytest.fixture
-def revaultd_factory(test_base_dir, bitcoind):
-    factory = RevaultDFactory(test_base_dir, bitcoind)
+def revaultd_factory(directory, bitcoind):
+    factory = RevaultDFactory(directory, bitcoind)
 
     yield factory
